consolidate to single store file setup
