+++ app/index.js
@@ -2,7 +2,7 @@ import React, { Fragment } from 'react';
 import { render } from 'react-dom';
 import { AppContainer as ReactHotAppContainer } from 'react-hot-loader';
 import Root from './containers/Root';
-import { configuredStore, history } from './store/configureStore';
+import { history, configuredStore } from './store';
 import './app.global.css';
 
 const store = configuredStore();
