+++ app/rootReducer.js
@@ -1,8 +1,8 @@
 // @flow
-import { combineReducers } from 'redux';
+import { combineReducers } from '@reduxjs/toolkit';
 import { connectRouter } from 'connected-react-router';
 import type { HashHistory } from 'history';
-import { counterReducer } from '../features/counter/counterSlice';
+import { counterReducer } from './features/counter/counterSlice';
 
 export default function createRootReducer(history: HashHistory) {
   return combineReducers<{}, *>({
