--- app/store/configureStore.js
@@ -1,12 +1,0 @@
-// @flow
-import configureStoreDev from './configureStore.dev';
-import configureStoreProd from './configureStore.prod';
-
-const selectedConfigureStore =
-  process.env.NODE_ENV === 'production'
-    ? configureStoreProd
-    : configureStoreDev;
-
-export const { configuredStore } = selectedConfigureStore;
-
-export const { history } = selectedConfigureStore;
